Update cydr.py
I streamed lined the SD card mount and unmount functions. Users no longer need to declare that they are using the SD card when they initialize an instance of the CDY class. Additional comments offering a brief list of the available functions are now listed at the beginning of the CDY class.